\name{acid}
\docType{data}
\alias{acid}
\title{Lake Acidity Study}
\description{The \code{acid} data frame has 112 rows and 4 columns of data derived based on the Eastern Lakes Survey of 1984 implemented by the Environmental Protection Agency of the USA.}

\usage{
data(acid)
}
\format{
The data frame contains the following columns:

ph a numeric vector of surface pH values.

t1 a numeric vector of calcium concentrations in log10 milligrams per liter.

x1, x2 numeric vectors of the lakes' geographic locations.
}

\details{112 lakes are extracted in the southern Blue Ridge mountains area. The surface pH values were recorded together with the calcium concentration and geographic locations.}

\source{
Douglas, A. and Delampady, M. (1990), Eastern Lake Survey Phase I: Documentation for the Data Base and the Derived Data sets. Tech Report 160 (SIMS), Dept. Statistics, University of British Columbia. 
}
\references{
Gu, C. and Wahba, G. (1993) Smoothing Spline ANOVA with component-wise Bayesian confidence intervals. Journal of Computational and Graphic Statistics 55, 353-368.
}
\keyword{datasets}
