\name{addCorrGroup}
\alias{addCorrGroup}
\title{Add Grouping Factor(s) to a Correlation Structure}
\description{
Retrun a correlation structure with the additional grouping factors added.
}
\usage{
addCorrGroup(cor.struct, cor.form)
}

\arguments{
  \item{cor.struct}{a correlation object of class \code{corStruct}.}
  \item{cor.form}{ a one side formula with grouping factors on the right side. }
}
\details{
  This function is used internally.
}
\value{
 Returned is a correlation object.
}
\keyword{file}
