\name{alogit}
\alias{alogit}
\title{ Calculate the Inverse Logit Transformation}
\description{
 Perform an inverse logit calculation
}
\usage{
alogit(x)
}
\arguments{
  \item{x}{a numeric value}
}

\value{
 Returned is \eqn{e^x/(1+e^x)}.
}
\author{Chunlei ke \email{chunlei\_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\keyword{file}


