\name{Arosa}
\docType{data}
\alias{Arosa}
\title{Monthly Mean Ozone Thickness in Arosa of Switzerland}
\description{ The \code{Arosa} data frame has 518 rows and 3 columns of data for monthly mean ozone thickness.}
\usage{
data{Arosa}
}
\format{
The data frame contains the following columns:

year a vector of integers from 1 to 46 indicating the years when the measures were taken from 1926. 

month a vector of integers from 1 to 12 represeting the months in a year.

thick a numeric vetor of mean ozone thickness (Dobson units).
}

\details{ Monthly mean ozone thickness in Arosa, Switzerland was recorded from 1926-1971.
}

\source{
Andrew,D. F. and Herzberg, A. M. (1985). Data: a collection of problems from many fields for the  
students and research workers. Springer: Berlin: New York.
}
\keyword{datasets}
