\name{canadaTemp}
\docType{data}
\alias{canadaTemp}
\title{Monthly Mean Temperatures}
\description{ The \code{canadaTemp} data frame has 420 rows and 3 columns of data for monthly mean temperatures in Canada}
\usage{
data(canadaTemp)
}
\format{
The data frame contains the following columns:

temp a numeric vector of mean temperatures at some stations in Canada. 

month a vector of integers from 1 to 12 represeting the months in a year.

station a vetor of integers from 1 to 35 indicating the sations where the temperatures were recorded.
}

\source{
The data set was downloaded from \url{http://www.psych.mcgill.ca/misc/fda/downloads/FDAfuns/R/data/}.
}
\references{
Ramsay, J. O and Silverman, B. W. (1997). Functional Data Analysis. New York:Springer.

Ke, C. and Wang, Y. (2001). Semi-parametric Nonlinear Mixed Effects Models and
Their Applications. JASA 96:1272-1298.
}
\keyword{datasets}

