\name{chickenpox}
\docType{data}
\alias{chickenpox}
\title{ Chickenpox in New York City}
\description{The \code{chickenpox} data frame has 498 rows and 3 columns of data 
recording the number of Chickenpox occurrences in New York City.}

\usage{
data(chickenpox)
}
\format{
The data frame contains the following columns:

count the number of monthly reported Chickenpox cases.

month a vector of integers from 1 to 12 representing the month for the reported cases.
year a numeric vector representing the year when the cases were reported.
}

\details{ This data frame contains monthly number of reported cases of chickenpox in New York City from 1931 to the first six months of 1972.}

\section{Source}{
The data were downloaded from \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL}.
}
\keyword{datasets}
