\name{climate}
\docType{data}
\alias{climate}
\title{Winter Average Temperatures}
\description{The data frame \code{climate}, obrained from the Carbon Dioxide Information and Analysis Center, has 690 rows and 5 columns of data representing station winter temperature measurements.}

\usage{
data(climate)
}
\format{
The data frame contains the following columns:

temp a numeric vector of temperatures in celsius.

lat, long numeric vectors identifying the lattitudes and longitudes of the stations in.

lat.degree, long.degree  numeric vectors identifying the lattitudes and longitudes of the stations in degree.
}

\details{The station winter average temperatures were the averages of the December, January and Febuary monthly average temperatures obtained from the Jones/Wigley data files obtainable from the CDIAC at Oak Ridge National Laboratory in the files ndp020r1/jonesnh.data.Z and ndp020r1/jonessh.dat.Z in the pbu directory at 128.219.24.36.
}

\source{
Jones, P., Wigley, T. and Briffa, K.. lobal and hemisphere temperature anaomalies-land and marine instrumental records. In T. Boden, D. Kaiser, R. Sepanski, and F. Stoss, editors, Trends '93: A Compendium of Data on Global Change, ORNL/CDIAC-65, pages  603-608, Oak Ridge, TN 1994. CDIAC, Oak Ridge National Laboratory.
}
\keyword{datasets}
