% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonMar.R
\name{CNN}
\alias{CNN}
\title{Convolutional Neural Network (CNN) with two hidden layers}
\usage{
CNN(X,y,X_new,plot.figure = FALSE)
}
\arguments{
\item{X}{A list of predictor matrices.}

\item{y}{Binary response variable.}

\item{X_new}{A list of new matrices at which predictions are to made.}

\item{plot.figure}{Option for plotting trajectory of accuracy over epochs.}
}
\value{
The returned object is a list of components.

\code{prob} - The estimated probabilities of the binary response at the predictors Xnew.

\code{class} - The estimated binary response at the predictors Xnew.

\code{history} - The trajectory of accuracy over epochs.

\code{acc} - The classification accuracy on test data.
}
\description{
Implement a CNN with two hidden layers and ReLu activation.
}
