% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{describe}
\usage{
describe(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{optional list of variables or other dplyr selection}
}
\value{
data frame with summary statistics
}
\description{
describe provides summary statistics of variables. Logicals and factors are converted to
numeric and for strings statistics of string length are provided.
}
\examples{
describe(cars)
}
