\name{dog}
\docType{data}
\alias{dog}
\title{Coronary Ainus Potassium Concentrations}
\description{The \code{dog} data frame has 252 rows and 4 columns of data considered by Grizzle and Alen (1969)
}
\usage{
data(dog)
}
\format{
The data frame contains the following columns:

y a numeric vector of meansurements of coronary sinus postassium concentrations.

group a vector of group index for the four groups of dogs.

dog a vector of integers identifying dogs.

time a numeric vector of time points measurements were made.
}

\details{The data are coronary sinus potassium concentrations measured on each of 36 dogs. These 36 dogs were divided into 4 treatment groups, and the measurements on each dog were taken every two minutes from 1 to 13 minutes after occlusion.}

\source{
Grizzle, J. E. and Allen, D. M. (1969). Analysis of growth and dose response curves, Biometrics 25: 357-381.
}
\keyword{datasets}


