\name{dsms}
\alias{dsms}
\title{Interface to Fortran Subroutine dsms }
\description{
  Calculate a matrix operation needed to construct Bayesian confidence intervals
}
\usage{
dsms(rkpk.obj)
}
\arguments{
  \item{rkpk.obj}{an object returned from calling dsidr}
}

\value{
a matrix. See the corresponding Fortran subroutine. 
}
\keyword{file}
