\name{getParaModelMatrix}
\alias{getParaModelMatrix}
\title{Get the Model Matrix for a Parameter Model}
\description{
Extract the model matrix for a model formula specified for parameter(s) 
}
\usage{
getParaModelMatrix(params, data, nobs)
}
\arguments{
  \item{params}{a two-sided formula}
  \item{data}{a data frame}
  \item{nobs}{an integer}
}
\details{
This function is used internally.
}
\keyword{file}

