\name{list.prod}
\alias{list.prod}
\title{Operation of Two Lists}
\description{
Perform compoentwise operation of two lists.
}
\usage{
list.prod(x, y, fun = "*")
}
\arguments{
  \item{x}{a numeric list}
  \item{y}{a numeric list of the same type as \code{x}}
  \item{fun}{a function. Default is product.}
}
\details{
This function is used internally.
}
\value{
a list.
}
\keyword{file}

