\name{paramecium}
\docType{data}
\alias{paramecium}            
\title{Growth of paramecium caudatum population}
\description{
The 'paramecium' data frame has 25 rows and 2 columns of data from an experiment that grow paramecium caudatum  
}
\usage{
data(paramecium)
}     
\section{Format}{
The data frame contains the following columns:

day a numeric vector of days since the start of the experiment

density a numeric vector of mean number of individuals in 0.5 ml of medium of four different cultures started simultaneously 
}
\section{Source}{
Gause, G.F. (1934). The Struggle for Existence. Baltimore, MD: Williams & Wilkins.
}
\section{references}{
Neal, D. (2004). Introduction to Population Biology. Cambridge University Press.
}
\keyword{datasets}

