\name{paste2}
\alias{paste2}
\title{A Special Paste Operation}
\description{
This function is used internally.
}
\usage{
paste2(nam, value)
}
\arguments{
  \item{nam}{a vector of strings.}
  \item{value}{a numeric vector with the same length as that of \code{nam}}
}

\value{
The retuened is a vector of the form (nam[1]=value[1], dots, nam[n]=value[n]).
}
\keyword{file}


