\name{ print.anova.ssr}
\alias{ print.anova.ssr}
\title{
Print an anova.ssr Object
}
\description{
Calculate and output p-values for tests available.
}
\usage{
\method{print}{anova.ssr}(x, ...)
}
\arguments{
 \item{x}{
an object inheriting from class anova.ssr, generally obtained by 
applying the anova.ssr method to an ssr object.
}
\item{\dots}{other available arguments, currently unused.}
}
\author{Chunlei Ke \email{chunlei_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
\code{\link{anova.ssr}}, \code{\link{ssr}}
}
\keyword{file}
