\name{print.nnr}
\alias{print.nnr}
\title{
Print Values
}
\description{
Print the arguments of a 'nnr' object.
}
\usage{
\method{print}{nnr}(x, ...)
}
\arguments{
\item{x}{a \code{nnr} object}
\item{\dots}{unused argument}
}
\details{
This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{nnr}. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{nnr}}
}
\keyword{file}

