\name{print.slm}
\alias{print.slm}
\title{
Print Values
}
\description{
Print the arguments of a \code{slm} object.
}

\usage{
print.slm(x, ...)
}
\arguments{
\item{x}{a \code{slm} object}
\item{\dots}{unused argument}
}
\details{
This is a method for  the  function  \code{print}  for  objects
       inheriting from class \code{slm}.  
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{slm}}
}
\keyword{file}

