\name{print.snm}
\alias{print.snm}
\title{
Print Values
}
\description{
Print the arguments of a 'snm' object.
}
\usage{
\method{print}{snm}(x, ...)
}
\arguments{
\item{x}{ a \code{snm} object}
\item{\dots}{unused argument}
}

\details{This is a method for  the  function \code{print}  for  objects
       inheriting from class `snm'. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{slm}}, \code{\link{print}}
}
\keyword{file}



