\name{print.snr}
\alias{print.snr}
\title{
Print Values
}
\description{
Print the arguments of a \code{snr} object.
}
\usage{
\method{print}{snr}(x, ...)
}
\arguments{
\item{x}{ a \code{snr} object}
\item{\dots}{unused argument}
}
\details{
This is a method for  the  function  \code{print}  for  objects
       inheriting from class \code{snr}. 
}
\author{Chunlei Ke \email{chunlei_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{snr}}
}
\keyword{file}

