\name{print.ssr}
\alias{print.ssr}
\title{
Print Values
}
\description{
Print the arguments of a \code{ssr} object.
}
\usage{
\method{print}{ssr}(x, ...)
}
\arguments{
  \item{x}{a \code{ssr} object}
  \item{\dots}{unused argument}
}
\details{
This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{ssr}. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{ssr}}
}
\keyword{file}
