\name{print.summary.nnr}
\alias{print.summary.nnr}
\title{Print Vales}
\description{
Print the arguments of a \code{summary.nnr} object
}
\usage{
\method{print}{summary.nnr}(x, ...)
}
\arguments{
  \item{x}{ an object of class \code{summary.nnr}}
  \item{\dots}{unused argument}
}
\details{
 This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{summary.nnr}. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{nnr}}, \code{\link{summary.nnr}}
}
\keyword{file}



