\name{print.summary.slm}
\alias{print.summary.slm}
\title{Print Values }
\description{
Print the arguments of a \code{summary.slm} object
}
\usage{
print.summary.slm(x, ...)
}
\arguments{
  \item{x}{ an object of class \code{summary.slm}}
  \item{\dots}{unused argument}
}
\details{
 This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{summary.slm}. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{slm}}, \code{\link{summary.slm}}
}
\keyword{file}



