\name{print.summary.snm}
\alias{print.summary.snm}
\title{Print Values }
\description{
Print the arguments of a \code{summary.snm} object
}
\usage{
\method{print}{summary.snm}(x, ...)
}
\arguments{
  \item{x}{ an object of class \code{summary.snm}}
  \item{\dots}{unused argument}
}
\details{
 This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{summary.snm}. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{snm}}, \code{\link{summary.snm}}
}
\keyword{file}



