\name{print.summary.snr}
\alias{print.summary.snr}
\title{Print Values }
\description{
Print the arguments of a \code{summary.snr} object
}
\usage{
\method{print}{summary.snr}(x, ...)
}
\arguments{
  \item{x}{ an object of class \code{summary.snr}}
  \item{\dots}{unused argument}
}
\details{
 This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{summary.snr}. 
}
\author{Chunlei Ke \email{chunlei\_ke@pstat.ucsb.edu} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{snr}}, \code{\link{summary.snr}}
}
\keyword{file}

