\name{print.summary.ssr}
\alias{print.summary.ssr}
\title{Print Values }
\description{
Print the arguments of a \code{summary.ssr} object
}
\usage{
\method{print}{summary.ssr}(x, ...)
}
\arguments{
  \item{x}{ an object of class \code{summary.ssr}}
  \item{\dots}{unused argument.}
}
\details{
 This is a method for  the  function \code{print}  for  objects
       inheriting from class \code{summary.ssr}. 
}
\author{Chunlei Ke \email{chunlei_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
   \code{\link{ssr}}, \code{\link{summary.ssr}}
}
\keyword{file}


