% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_join.R
\name{safe_join}
\alias{safe_join}
\title{safe left join}
\usage{
safe_join(x, y, by = NULL, return_errors = FALSE, ...)
}
\arguments{
\item{x}{Data frame}

\item{y}{Data frame}

\item{by}{Key(s) to join by. Required.}

\item{return_errors}{Set to TRUE to return error dataset}

\item{...}{Other parameters sent to dplyr}
}
\value{
Joined dataset
}
\description{
safe_join performs a dplyr left_join, checking for common errors and warning
when matches are not found.
}
\details{
safe_join performs the following checks:

1. Checks that key is unique in df2

2. Reports that matches not found in df1 or df2 (Corresponding to _merge == 1)

3. Checks if non-key columns exist in both datasets
}
