\name{star}
\docType{data}
\alias{star}
\title{           
Magnitude of the Mira Variable R Hydrae
}
\description{
The \code{star} data frame has 1086 rows and 2 columns of data from the Mira Variable R Hydrae
}
\usage{
data(star)
}     
\section{Format}{
The data frame contains the following columns:

time a numeric vector of the observation time in days

magnitude a numeric vector of brightness of the Mira variable R Hydrae
}

\details{
This dataset contains magnitude (brightness) of the Mira variable R Hydrae during 1900-1950.
}
\section{Source}{
Genton, M. G. and Hall, P. (2007). Statistical Inference for Envolving Periodic Functions, Journal of the Royal Statistical Society B 69, 643-657.
}
\section{references}{
Yuedong Wang and Chunlei Ke (2009), Smoothing Spline Semi-parametric Nonlinear Regression Models, Journal of Computational and Graphical Statistics 18, 165-183.
}
\keyword{datasets}

