\name{Stratford}
\docType{data}
\alias{Stratford}            
\title{Daily maximum temperatures in Stratford}
\description{
The \code{Stratford} data frame has 73 rows and 2 columns of data containing daily maximum temperatures in Stratford every five days in 1990
}
\usage{
data(Stratford)
}     
\section{Format}{
The data frame contains the following columns:

x a numeric vector representing time in a year scaled into [0,1]

y a numeric vector of the observed maximum temperature in Fahrenheit  
}
\details{
Daily maximum temperatures from the station in Stratford, Texas, in the year 1990 were extracted. 
The year was divided into 73 five-day periods and measurements on the third day in each period were selected as observations. 
}
\section{Source}{
This is part of a climate dataset downloaded from the Carbon Dioxide Information Analysis Center at http://cdiac.ornl.gov/ftp/ndp070. 
}
\keyword{datasets}
