\name{sumList}
\alias{sumList}
\title{Sum of Components of a List}
\description{
This function is used internally.
}
\usage{
sumList(x)
}
\arguments{
  \item{x}{a numeric list with the same structure for all components.}
}
\value{
Returned is summation of all components.
}
\keyword{file}


