\name{summary.nnr}
\alias{summary.nnr}
\title{
Object Summaries
}
\description{
Summarize a \code{nnr} object
}
\usage{
\method{summary}{nnr}(object, ...)
}
\arguments{
\item{object}{a fitted \code{nnr} object.}
\item{\dots}{unused argument}
}
\details{
This is a method for the function  \code{summary}  for  objects inheriting from class
 \code{nnr}.  See summary for the general behavior of this function.
}
\author{Chunlei Ke \email{chunlei\_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
\code{\link{nnr}}, \code{\link{print.nnr}}
}
\keyword{file}

