\name{summary.slm}
\alias{summary.slm}
\title{
Object Summaries
}
\description{
Summarize a \code{slm} object
}
\usage{
\method{summary}{slm}(object, ...)
}
\arguments{
\item{object}{a fitted \code{slm} object.}
\item{\dots}{unused argument}
}
\details{
This is a method for the function  \code{summary}  for  objects inheriting 
from class \code{slm}.
}
\author{Chunlei Ke \email{chunlei_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
\code{\link{slm}}, \code{\link{print.slm}}
}
\keyword{file}

