\name{summary.snm}
\alias{summary.snm}
\title{
Object Summaries
}
\description{
Summarize a \code{snm} object
}
\usage{
summary.snm(object, ...)
}
\arguments{
\item{object}{a fitted 'snm' object.}
\item{\dots}{unused argument}
}
\details{
This is a method for the function  \code{summary}  for  objects inheriting from class
 \code{snm}. 
}
\author{Chunlei Ke \email{chunlei\_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
\code{\link{snm}}, \code{\link{print.snm}}
}
\keyword{file}
