\name{summary.snr}
\alias{summary.snr}
\title{
Object Summaries
}
\description{
Summarize a 'snr' object
}
\usage{
\method{summary}{snr}(object, ...)
}
\arguments{
\item{object}{a fitted \code{snr} object.}
\item{\dots}{unused argument}
}
\details{
This is a method for the function  \code{summary}  for  objects inheriting from class
 \code{snr}.  See summary for the general behavior of this function.
}
\author{Chunlei Ke \email{chunlei\_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
\code{\link{snr}}, \code{\link{print.snr}}
}
\keyword{file}


