\name{summary.ssr}
\alias{summary.ssr}
\title{
Summarize a ssr object
}
\description{
Provides a synopsis of a \code{ssr} object and perform tests.
}
\usage{
summary.ssr(object, ...)
}
\arguments{
  \item{object}{ a fitted \code{ssr} object. }
  \item{\dots}{unused option.}
}
\details{
This is a method for the function  \code{summary}  for  objects inheriting 
from class \code{ssr}. 
}
\author{Chunlei Ke \email{chunlei\_ke@yahoo.com} and Yuedong Wang \email{yuedong@pstat.ucsb.edu}}
\seealso{
\code{\link{ssr}}, \code{\link{print.ssr}}
}
\keyword{file}

