\name{table2}
\alias{table2}
\title{Extension of the Table Function }
\description{
 This function is used internally.
}
\usage{
table2(..., exclude = c(NA, NaN))
}
\arguments{
  \item{\dots}{as described in \code{table} }
  \item{exclude}{as described in \code{table}}
}
\value{
 a table.
}
\keyword{file}

