\name{tp.term}
\alias{tp.term}
\title{Intermediate Steps to Computer Thin Plate Spline Kernel}
\description{
This function is used internally.
}
\usage{
tp.term(x, order = 2)
}
\arguments{
\item{x}{
a list or matrix of observations. One component, if a list, and one column, if a matrix,
contains observations on one variable. If a list, all components must be of the same length. 
}
\item{order}{
an optional integer specifying the order of the thin plate spline. Default is 2.
}}

\value{
A matrix
}
\keyword{file}


