\name{USAtemp}
\docType{data}
\alias{USAtemp}
\title{Average Winter temperature in the United States}
\description{
The \code{USAtemp} data frame has 1214 rows and 3 columns of data containing average Winter temperatures in 1981 from 1205 stations in USA.
}
\usage{
data(USAtemp)
}
\section{Format}{
The data frame contains the following columns:

temp a numeric vector of average temperatures (Fahrenheit)

lat a numeric vector of the latitude of a station       

long a numeric vector of the longitude of a station       
}
\section{details}{
The average Winter temperatures are calculated as the averages of temperatures in December, January and February.
The geological locations of 1214 stations are given in terms of longitude and latitude.
}

\keyword{datasets}
