\name{wesdr}
\docType{data}
\alias{wesdr}
\title{Wisconsin Epidemiological Study of Diabetic Retinopathy}
\description{The \code{wesdr} data frame has 669 rows and 5 columns of data
from an ongoing epidemiological study of a cohort of patients receiving their medical care in an 11-country area in southern Wisconsin.}

\usage{
data(wesdr)
}
\section{Format}{
This data frame contains the following columns:

num a numeric vector giving IDs for individuals.

dur a numeric vector of duration of at baseline in year.

gly a numeric vector of glycosylated hemoglobin, a measuer of hyperglycemia.

bmi a numeric vecttor of body mass index, weight in \eqn{kg/(height in meter)^2}.

prg a vector of 0 or 1's representing disease progression for each individual.
}

\details{ The progression of diabetic retinopathy was assessed together with a number of medical, demographic, ocular and other covariates and the retinopathy scores.}

\section{Source}{
Klein, R., Klein, B. E. K., Moss, S. E., Davis, M. D. and Demets, D. L. (1989a). The Wisconsin epidemiologic study of diabetic retinopathy. IX. Four year incidence and progression of diabetic retinopathy when age at diagnosis is less than 30 years. Arch. Ophthalmal. 107, 237-243.

Klein, R., Klein, B. E. K., Moss, S. E., Davis, M. D. and Demets, D. L. (1989b). The Wisconsin epidemiologic study of diabetic retinopathy. X. Four year incidence and progression of diabetic retinopathy when age at diagnosis is less than 30 years. Arch. Ophthalmal. 107, 244-249.
}
\keyword{datasets}

